# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ProЗапас API",
    "version": "1.0.0",
    "description": "HTTP-интерфейс между десктопным приложением и сервером.\n\nСистема обслуживает несколько складов одной компании. Каждый пользователь привязан к одному складу, и `warehouse_id` берётся из токена — в запросах он не передаётся. Сервер сам подставляет фильтр видимости, поэтому чужие заказы недостижимы.\n\nЗаказ — перемещение товара между двумя складами. Направление не хранится: для склада-заказчика заказ исходящий, для отправителя входящий.\n\nПодробности и обоснования — docs/api.md."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/catalog": {
      "get": {
        "tags": [
          "Справочник"
        ],
        "summary": "Номенклатура",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "total",
                    "limit",
                    "offset"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CatalogItem"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 128
                    },
                    "limit": {
                      "type": "integer",
                      "example": 50
                    },
                    "offset": {
                      "type": "integer",
                      "example": 0
                    }
                  },
                  "description": "Страница справочника"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          }
        },
        "description": "Поиск q идёт по генерируемой колонке: наименование, артикул, код 1С.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 200
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ]
      },
      "post": {
        "tags": [
          "Справочник"
        ],
        "summary": "Добавить позицию",
        "responses": {
          "201": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItem"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogCreate"
              }
            }
          }
        }
      }
    },
    "/catalog/{id}": {
      "get": {
        "tags": [
          "Справочник"
        ],
        "summary": "Карточка позиции",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItem"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      },
      "patch": {
        "tags": [
          "Справочник"
        ],
        "summary": "Изменить позицию",
        "responses": {
          "200": {
            "description": "Успех",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItem"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogUpdate"
              }
            }
          }
        }
      }
    },
    "/catalog/{id}/archive": {
      "post": {
        "tags": [
          "Справочник"
        ],
        "summary": "Архивировать позицию",
        "responses": {
          "204": {
            "description": "Готово"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          }
        },
        "description": "Позиция с движениями не удаляется, а архивируется.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access-токен. Внутри — склад пользователя и права."
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "description": "RFC 9457",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "example": "https://prozapas/errors/insufficient-stock"
          },
          "title": {
            "type": "string",
            "example": "Недостаточно свободного остатка"
          },
          "status": {
            "type": "integer",
            "example": 409
          },
          "detail": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
...
      "CatalogItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "article": {
            "type": "string"
          },
          "code1c": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "unit_weight": {
            "type": "number"
          },
          "is_archived": {
            "type": "boolean"
          }
        }
      },
      "CatalogCreate": {
        "type": "object",
        "required": [
          "article",
          "name",
          "unit"
        ],
        "properties": {
          "article": {
            "type": "string"
          },
          "code1c": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "unit_weight": {
            "type": "number",
            "default": 0
          }
        }
      },
      "CatalogUpdate": {
        "type": "object",
        "properties": {
          "code1c": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "unit_weight": {
            "type": "number"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Тело или параметры не разобрались",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Нет токена, истёк или отозван",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Нет прав на раздел или запрещено правилом",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "NotFound": {
        "description": "Объекта нет, удалён или принадлежит чужим складам",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Conflict": {
        "description": "Конфликт версий, недопустимый переход, нехватка остатка",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Unprocessable": {
        "description": "Нарушено доменное правило",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
    }
  }
}